feat: add loopwordswitcher component
