+++ .gitignore
@@ -22,6 +22,7 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 .yarnrc.yml
+.yarn
 .eslintcache
 package-lock.json
 pnpm-lock.yaml
